+++ package.json
@@ -22,7 +22,11 @@
   },
   "homepage": "https://github.com/marcode95/to-do-list#readme",
   "devDependencies": {
+    "babel-eslint": "10.1.x",
     "css-loader": "^5.2.0",
+    "eslint": "6.8.x",
+    "eslint-config-airbnb-base": "14.1.x",
+    "eslint-plugin-import": "2.20.x",
     "style-loader": "^2.0.0",
     "webpack": "^5.30.0",
     "webpack-cli": "^4.6.0"
